Moving the common code to shared folder
